+++ package.json
@@ -9,11 +9,14 @@
     "test": "jest --config jest.config.cjs",
     "test:integration": "jest --config jest.config.integration.cjs",
     "start": "npm run serve",
+    "start:auth": "npm run serve:auth",
     "tsc": "./node_modules/typescript/bin/tsc",
     "build": "tsc",
     "serve": "node dist/server.js",
+    "serve:auth": "node dist/authServer.js",
     "postinstall": "npm run build",
-    "dev": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/server.ts"
+    "dev": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/server.ts",
+    "dev:auth": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/authServer.ts"
   },
   "keywords": [],
   "author": "",
